Use single loader to preserve lines
